gptinder_back/requirements.txt
Django
djangorestframework
django-cors-headers
python-dotenv
openai
numpy
boto3
dj-database-url
django-celery-beat
django-filter
django-redis
django-storages
psycopg2-binary
redis
pyTelegramBotAPI
drf-yasg
gunicorn
pillow
pinecone